% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pcpn_isotopes.R
\name{add_pcpn_isotopes}
\alias{add_pcpn_isotopes}
\title{Add precipitation isotopes to dataset}
\usage{
add_pcpn_isotopes(tracer, start_date, end_date)
}
\arguments{
\item{tracer}{data frame with water chemistry information for stable isotope.}

\item{start_date}{start date of analysis period, POSIXct}

\item{end_date}{end date of analysis period, POSIXct}
}
\value{
tracer, same data frame as provided, but with additional
  precipitation measurements during the analysis period for months with
  missing precipitation measurements.
}
\description{
Adds precipitation stable isotope measurements from Maribeth Kniffin (May
2016 - April 2017) and incorporates them in the current timeseries, but only
for months without CSLS precipitation measurements.
}
