% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_C_GWin.R
\name{calculate_C_GWin}
\alias{calculate_C_GWin}
\title{Calculate lake concentration of parameter}
\usage{
calculate_C_GWin(
  C_vals,
  fluxes = CSLSdata::MODFLOW,
  scenario = "cal",
  sim = 0,
  start_date = as_datetime("2017-10-01"),
  end_date = as_datetime("2018-09-30"),
  lakes = c("Pleasant", "Long", "Plainfield")
)
}
\arguments{
\item{C_vals}{data frame with columns for "lake", "C_pcpn", "C1_lake",
"C2_lake", and "C_GWout"}

\item{fluxes}{data frame with lake, date, level_m, GWin_m3, and P_m3 from
MODFLOW simulations to use for mass balance calculations.}

\item{scenario}{which MODFLOW scenario is being evaluated here, defaults to "cal".}

\item{sim}{which MODFLOW simulation is being evaluated here, defaults to 0}

\item{start_date}{start date for evaluation, defaults to 2017-10-01}

\item{end_date}{end date for evaluation, defaults to 2018-09-30.}

\item{lakes}{name of lakes to include in analysis, defaults to
c("Pleasant", "Long", "Plainfield")}
}
\value{
C_GWin, a data frame with the following columns:
  \item{lake}{name of lakes, i.e., "Pleasant", "Long", and "Plainfield"}
  \item{C_GWin}{calculated concentration in groundwater inflow}
}
\description{
Calculates the concentration of a parameter in the lake given an initial lake
concentration, monthly fluxes of precipitation, groundwater, and lake volume,
and median concentrations of groundwater and precipitation.
}
