% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ice_thickness.R
\name{calculate_ice_thickness}
\alias{calculate_ice_thickness}
\title{Calculate ice thickness}
\usage{
calculate_ice_thickness(
  date,
  ice_on = 320,
  ice_off = 105,
  ice_max = 60,
  ice_max_depth = NISTunits::NISTinchTOmeter(24) * 1000
)
}
\arguments{
\item{date}{date to calculate ice thickness (POSIX)}

\item{ice_on}{day of year when ice starts forming. Defaults to day 320 (Nov
16).}

\item{ice_off}{day of year when ice is all gone. Defaults to day 105 (Apr 15)}

\item{ice_max}{day of year when ice thickness is at a maxiumum. Defaults to
day 60 (Mar 1).}

\item{ice_max_depth}{max thickness of ice (m). Defaults to 24 inches.}
}
\value{
ice_depth, a numeric value of ice thickness (m).
}
\description{
Calculates the thickness of ice on a lake. Assumes a linear increase from a
static ice-on date to a static ice-off date with a static date of max
thickness and a static max thickness.
}
