% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_water_balance_tracer.R
\name{calculate_water_balance_tracer}
\alias{calculate_water_balance_tracer}
\title{Calculate water balance}
\usage{
calculate_water_balance_tracer(
  param = "d18O",
  start_date = as_datetime("2018-10-01"),
  end_date = as_datetime("2019-09-30"),
  dt = "annual",
  no_ice = FALSE,
  C_evap = 0,
  C_ice = 0,
  mean_lake = TRUE
)
}
\arguments{
\item{param}{name of water chemistry parameter to grab concentration
information for. Must match the "description" field of
CSLSdata::water_chem for the desired parameter. Defaults to
"d18O". Options include "d18O", "d2H", "CALCIUM TOTAL
RECOVERABLE", "MAGNESIUM TOTAL RECOVERABLE", "CHLORIDE", and
"SODIUM TOTAL RECOVERABLE".}

\item{start_date}{start date of analysis. Defaults to start of WY2019
("2018-10-01").}

\item{end_date}{end date of analysis. Defaults to end of WY2019
("2019-09-30").}

\item{dt}{desired time step of inputs (e.g., "day" or "month")}

\item{no_ice}{logical defaults to TRUE to ignore ice formation.}

\item{C_evap}{concentraion in evaporation, defaults to zero}

\item{C_ice}{concentration in ice, defaults to zero.}

\item{mean_lake}{logical defaults to TRUE to calculate and use mean
of min and max lake value (should equate to fall and spring
turnover samples) for water budget calculations.}
}
\value{
lake_inputs
}
\description{
Calculates the water balance using desired tracer, e.g., stable isotopes or
conservative anion/cation.
}
