% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cevap_functions.R
\name{Cevap_normalized_humidity}
\alias{Cevap_normalized_humidity}
\title{Normalized Relative Humidity}
\usage{
Cevap_normalized_humidity(RH, es_a, es_l)
}
\arguments{
\item{RH}{relative humidity (percent)}

\item{es_a}{saturation vapor pressure for the air (kPa)}

\item{es_l}{saturation vapor pressure for the lake (kPa)}
}
\value{
h - the relative humidity normalized to the temperature of the
        surface water (-)
}
\description{
Calculates the relative humidity normalized to the temperature of the surface
water. Based on Equation 1.8 of Mook (2000).
}
\references{
Mook, W.G. (ed.) 2000. Environmental Isotopes in the Hydrologic
  Cycle: Volume III: Surface Water. UNESCO. Paris, France.
}
