% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_parameter.R
\name{filter_parameter}
\alias{filter_parameter}
\title{Filter SWIMS by parameter}
\usage{
filter_parameter(
  water_chem = CSLSdata::water_chem,
  param,
  plotting_name = "",
  numeric = TRUE,
  no_blanks = TRUE,
  no_dups = TRUE,
  no_age = FALSE,
  no_bad_well = FALSE,
  no_bad_sample = TRUE,
  no_lod = FALSE,
  no_comment = FALSE,
  note_lake_bottom = FALSE
)
}
\arguments{
\item{water_chem}{data frame with water_chem information for sites of interest}

\item{param}{parameter to subset by, must exactly match DNR description
(e.g., "ALUMINUM,TOTAL RECOVERABLE"). Can be a vector of
multiple descriptors.}

\item{plotting_name}{- name of parameter to use for plotting, e.g., "Total
Recoverable Alumninum", no units. Defaults to "".}

\item{numeric}{logical defaults to TRUE to indicate results should be numeric}

\item{no_blanks}{logical defaults to TRUE to exclude blank samples}

\item{no_dups}{logical defaults to TRUE to exclude duplicate samples}

\item{no_age}{logical defaults to FALSE to include samples that were analyzed
past the holding date.}

\item{no_bad_well}{logical defaults to FALSE to include samples from wells
known to behave weirdly.}

\item{no_bad_sample}{logical defaults to TRUE to exclude bad measurements
from QC.}

\item{no_lod}{logical defaults to FALSE to include samples with results below
the limit of detection.}

\item{no_comment}{logical defaults to FALSE to include samples with other
comments from the lab}

\item{note_lake_bottom}{logical defaults to FALSE to leave all lake samples
as "lake". If true, notes samples > 5m deep as
"lake_bottom".}
}
\value{
df, same data frame, with only the results for the
        parameter of interest. Also includes the given plotting name ("name").
}
\description{
Subsets SWIMS dataset for a CSLS lake to only measurements for a parameter of
interest.
}
