% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_C_lake.R
\name{get_C_lake}
\alias{get_C_lake}
\title{Calculate lake concentration}
\usage{
get_C_lake(
  param = "MAGNESIUM TOTAL RECOVERABLE",
  water_chem = CSLSdata::water_chem,
  start_date = NULL,
  end_date = NULL,
  summary_type = "median",
  value_name = "C_lake"
)
}
\arguments{
\item{param}{description of parameter to evaluate, defaults to
"MAGNESIUM TOTAL RECOVERABLE"}

\item{water_chem}{data frame with water chemistry to evaluate in
"filter_parameter", defaults toCSLSdata::water_chem.}

\item{start_date}{start date for evaluation, defaults to NULL to start with
first sample.}

\item{end_date}{end date for evaluation, defaults to NULL to end with last
sample.}

\item{summary_type}{Either "median", "min", or "max". Defaults to "median".}

\item{value_name}{name of column with resulting values. Defaults to "C_lake".}
}
\value{
C_pcpn, a data frame with the following columns:
  \item{lake}{name of lakes, i.e., "Pleasant", "Long", and "Plainfield"}
  \item{C_lake}{summary value of parameter in lake to use}
}
\description{
Calculates the concentration of a parameter in the lake across
all samples (default) or between given dates.
}
