% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_C_pcpn.R
\name{get_C_pcpn}
\alias{get_C_pcpn}
\title{Calculate precipitation concentration}
\usage{
get_C_pcpn(
  param = "MAGNESIUM TOTAL RECOVERABLE",
  water_chem = CSLSdata::water_chem,
  start_date = NULL,
  end_date = NULL,
  lakes = c("Pleasant", "Long", "Plainfield")
)
}
\arguments{
\item{param}{description of parameter to evaluate, defaults to
"MAGNESIUM TOTAL RECOVERABLE"}

\item{water_chem}{data frame with water chemistry to evaluate in
"filter_parameter", defaults toCSLSdata::water_chem.}

\item{start_date}{start date for evaluation, defaults to NULL to start with
first sample.}

\item{end_date}{end date for evaluation, defaults to NULL to end with last
sample.}

\item{lakes}{lakes to map precipitation values to.}
}
\value{
C_pcpn, a data frame with the following columns:
  \item{lake}{name of lake, i.e., "Pleasant", "Long", and "Plainfield"}
  \item{C_pcpn}{median value of parameter in precipitation to use}
}
\description{
Calculates the median concentration of a parameter in precipitation across
all samples (default) or between given dates.
}
