% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_lake_temp.R
\name{process_lake_temp}
\alias{process_lake_temp}
\title{Process lake temperature data for calcs}
\usage{
process_lake_temp(start_date, end_date, dt = "day")
}
\arguments{
\item{start_date}{start date of analysis period (POSIX).}

\item{end_date}{end date of analysis period (POSIX).}

\item{dt}{time step at which to summarize data. Defaults to "day" for daily
time step, can also be "month" for monthly time step.}
}
\value{
df, a data frame with the following columns:
  \item{lake}{name of lake, i.e., "Pleasant", "Long", and "Plainfield"}
  \item{date}{date of observation (POSIX). If monthly time step, monthly
              summary is assigned to first day of the month}
  \item{ltmp_bot_C}{lake temperature at bottom of lake, deg C. Used for some
                    solute reaction rates.}
  \item{ltmp_surf_C}{lake temperature at surface of lake, deg C. Used for
                     stable isotope in evaporation calculations}
}
\description{
Processes CSLS lake temperature data for water/solute balance calculations.
Filters HOBO temperature data from CSLSdata::water_chem for surface and
bottom temperatures, and summarizes at a daily or monthly time step for the
desired time period.
}
\details{
If using with non-CSLS data, will need an alternate way to get weather data
(including lake evaporation) into the format outputted by this function.
}
