% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_levels.R
\name{process_levels}
\alias{process_levels}
\title{Process lake elevation, area, volume data for calcs}
\usage{
process_levels(start_date, end_date, dt = "month")
}
\arguments{
\item{start_date}{start date of analysis period (POSIX).}

\item{end_date}{end date of analysis period (POSIX).}

\item{dt}{time step at which to summarize data. Defaults to "day" for daily
time step, can also be "month" for monthly time step.}
}
\value{
df, a data frame with the following columns:
  \item{lake}{name of lake, i.e., "Pleasant", "Long", and "Plainfield"}
  \item{date}{date of observation (POSIX). If monthly time step, monthly
              summary is assigned to first day of the month}
  \item{level_m}{mean lake elevation (mamsl) for the time step (daily or
                 monthly)}
  \item{area_m2}{mean lake area (m2) for the time step (daily or monthly)}
  \item{vol_m3}{mean lake volume (m3) for the time step (daily or monthly)}
  \item{dV_m3}{change in lake volume (m3) for the time step (daily or
               monthly)}
}
\description{
Processes CSLS lake level data for water/solute balance calculations. Ensures
no gaps in daily lake elevation data from CSLSdata::lake_levels, uses
CSLSdata::bathymetry to convert daily elevations to areas and volumes,
then summarizes to desired time step (daily or monthly).
}
\details{
If using with non-CSLS data, will need an alternate way to get lake level,
lake area, and lake volume data into the format outputted by this function.
}
