% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tracer.R
\name{process_tracer}
\alias{process_tracer}
\title{Process tracer data for water balance calcs}
\usage{
process_tracer(
  parameter,
  start_date,
  end_date,
  dt = "day",
  use_kniffin = TRUE,
  median_gw = TRUE,
  mean_lake = FALSE
)
}
\arguments{
\item{parameter}{description of parameter to use in analysis, as in
CSLSdata::water_chem$description (e.g., "d18O",
"MAGNESIUM TOTAL RECOVERABLE")}

\item{start_date}{start date of analysis period (POSIX).}

\item{end_date}{end date of analysis period (POSIX).}

\item{dt}{time step at which to summarize data. Defaults to "day" for daily
time step, can also be "month" for monthly time step.}

\item{use_kniffin}{logical defaults to TRUE to use kniffin precip stable
isotope data (only when stable isotopes selected as the
paramter)}

\item{median_gw}{logical defaults to TRUE to use median groundwater values,
rather than dynamic time series of interpolated groundwater
values.}

\item{mean_lake}{logical defaults to FALSE. If TRUE, calculate and use mean
of min and max lake value (should equate to fall and spring
turnover samples)}
}
\value{
df, a data frame with the following columns:
  \item{lake}{name of lake, i.e., "Pleasant", "Long", and "Plainfield"}
  \item{date}{date of observation (POSIX). If monthly time step, monthly
              means are assigned to first day of the month}
  \item{C_lake}{concentration in the lake (units in per mil, mg/L, or
                whatever unit parameter has in CSLSdata::water_chem)}
  \item{C_pcpn}{concentration in precipitation (units in per mil, mg/L, or
                whatever unit parameter has in CSLSdata::water_chem)}
  \item{C_GWin}{concentration in upgradient groundwater (units in per mil,
                mg/L, or whatever unit parameter has in
                CSLSdata::water_chem)}
}
\description{
Processes concentrations of parameters (e.g., d18O, Magnesium, Calcium) for
use in water/solute balances. Filters CSLSdata::water_chem to specified
parameter, limits to lake, precipitation, and upgradient groundwater sites,
and interpolates to a daily timestep.
}
\details{
If using with non-CSLS data, will need an alternate way to get solute/isotope
concentrations into the format outputted by this function.
}
