% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_weather.R
\name{process_weather}
\alias{process_weather}
\title{Process weather data for calcs}
\usage{
process_weather(start_date, end_date, dt = "day")
}
\arguments{
\item{start_date}{start date of analysis period (POSIX).}

\item{end_date}{end date of analysis period (POSIX).}

\item{dt}{time step at which to summarize data. Defaults to "day" for daily
time step, can also be "month" for monthly time step.}
}
\value{
df, a data frame with the following columns:
  \item{lake}{name of lake, i.e., "Pleasant", "Long", and "Plainfield"}
  \item{date}{date of observation (POSIX). If monthly time step, monthly
              summary is assigned to first day of the month}
  \item{day}{day of year of observation (1-366)}
  \item{P_mm}{precipitation (mm)}
  \item{E_mm}{lake evaporation (mm)}
  \item{atmp_C}{mean air temperature (deg C)}
  \item{RH_pct}{mean relative humidity (percent)}
  \item{irr_factor}{irradiance factor, fraction from 0-1 representing
                    relative intensity of solar radiation based on day of
                    year. For use in dynamic lake model for some solute
                    reactions}
}
\description{
Processes CSLS weather data for water/solute balance calculations. Transforms
data from CSLSdata package, calculates lake evaporation using CSLSevap
package, and summarizes at a daily or monthly time step for the desired time
period.
}
\details{
If using with non-CSLS data, will need an alternate way to get weather data
(including lake evaporation) into the format outputted by this function.
}
